// astr-ai-code-mother-frontend/lib/api/generated/model/appQueryRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface AppQueryRequest {
  pageNum?: number;
  pageSize?: number;
  sortField?: string;
  sortOrder?: string;
  id?: number;
  appName?: string;
  cover?: string;
  initPrompt?: string;
  userId?: number;
  codeGenType?: string;
  deployKey?: string;
  priority?: number;
}